requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
forex-python==1.9.2
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.2.0
propcache==0.3.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.4
simplejson==3.20.1
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.5.0
yarl==1.19.0
